Added client receiving messages properly

+++ Client/src/Game/clientUtils.java
@@ -1,10 +1,10 @@
 package Game;
 
 public class clientUtils {
-    public static void sleep(int secs)
+    public static void sleep(double secs)
     {
         try {
-            Thread.sleep(secs * 1000);
+            Thread.sleep((long) (secs * 1000));
         } catch (InterruptedException e) {
             System.out.println("Fuck");
             e.printStackTrace();
